+++ RELEASE_NOTES.md
@@ -1,4 +1,13 @@
-Bugfix release
+Improvement and Bugfix release. ATTENTION: The installation procedure has changed. Please check the updated doc/UPGRADEINFO.md file.
 
 **Bugfixes**
-- Fix bug in basic.sql ([Issue](https://github.com/glewe/teamcal-neo/issues/18))
+- Fix edit group calendar bug ([Issue](https://github.com/glewe/teamcal-neo/issues/19))
+
+**Improvements**
+- Improve group memebership edit behavior ([Issue](https://github.com/glewe/teamcal-neo/issues/20))
+- Improved installation and upgrade procedure ([Issue](https://github.com/glewe/teamcal-neo/issues/21))
+- Update dependency modules
+- Update summernote addon to 0.9.1
+- Update datatables addon to 2.3.8
+- Update Font Awesome to 7.2.0
+- Update jQuery to 4.0.0